Add Rust bindings
This is a Rust binding for the Lua binding, if that makes sense.  You'll
access tree-sitter parse trees using Lua code, but you'll manage that
Lua environment from Rust code.  I've chosen the `lua` crate because it
provides a Rust API that most closely mimics the actual Lua C API.